Migrate the glib based test case for GncTxImport to google-test
