+++ src/import-export/csv-imp/test/test-csv-imp.c
@@ -32,7 +32,6 @@
  * each sub-suite; avoids having header files. */
 
 extern GTestSuite *test_suite_gnc_csv_model();
-extern GTestSuite *test_suite_gnc_csv_imp_trans();
 
 int
 main (int   argc,
@@ -56,7 +55,6 @@ main (int   argc,
      * details. Unfortunately, GLib-Testing doesn't provide the automatic
      * registration features of more sophisticated frameworks. */
     test_suite_gnc_csv_model();
-    test_suite_gnc_csv_imp_trans();
 
     return g_test_run();
 }
